Change to schema

+++ Database_setup.sql
@@ -6,7 +6,7 @@ CREATE TABLE stockdemo (
 	transaction_id BIGINT auto_increment PRIMARY KEY,
     ticker VARCHAR(50) NOT NULL,
     stock_name VARCHAR(50) NOT NULL,
-    stock_value DECIMAL NOT NULL,
-    quantity DECIMAL NOT NULL,
+    stock_value float NOT NULL,
+    quantity float NOT NULL,
     transaction_date DATE DEFAULT (now())
     );
